pyproject.toml
[tool.poetry]
name = "release-pr-bot"
version = "0.1.0"
description = ""
authors = ["Paul Gessinger <hello@paulgessinger.com>"]
packages = [
    { include = "release_pr_bot", from = "src" },
]

[tool.poetry.dependencies]
python = "^3.8"
sanic = "^20.6.3"
gidgethub = "^4.1.1"
aiohttp = "^3.6.2"
python-dotenv = "^0.14.0"
cachetools = "^4.1.1"
python-semantic-release = "^7.2.2"

[tool.poetry.dev-dependencies]
pysmee = "^0.1.1"
pytest = "^6.0.1"
black = "19.10b0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
